Restructure tests and move mocks out to test-lib/mocks
