+++ platformio.ini
@@ -129,12 +129,14 @@ build_unflags =
 
 [env:test]
 platform = native
-build_flags =
-	-std=gnu++17
-	-g -O0
+build_flags = -std=gnu++11
 lib_deps =
+	mocks
 	ArduinoFake
-	powerbroker2/SerialTransfer
+	powerbroker2/SerialTransfer@^3.1.3
+lib_extra_dirs =
+	test-lib
+lib_compat_mode = off
 
 [env:webserver]
 platform = https://github.com/platformio/platform-espressif32.git
